[Bug fix] Index out of bounds in builder.rs
The build method in TyVcdBuilder wasn't getting the kind of a vector properly.
The previous tests were working for a coincidence, indeed the issue occurs only when a vector of bundle is used and its type contains more than 2 fields (n_fields > 2).
But now, that test case is covered and the issue is solved.